requirements.txt
addict==2.4.0
aenum==3.1.16
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anthropic==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astroid==4.0.2
attrs==25.4.0
audioread==3.1.0
bcrypt==5.0.0
cachetools==6.2.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.3
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
dashscope==1.25.3
dataclasses-json==0.6.7
datasets==2.18.0
decorator==5.2.1
deepgram-sdk==3.7.7
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
dnspython==2.8.0
edge-tts==7.2.3
editdistance==0.8.1
einops==0.8.1
email-validator==2.3.0
fastapi==0.104.1
filelock==3.20.0
flake8==7.3.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2024.2.0
funasr==1.1.12
gast==0.7.0
gotrue==2.12.4
gTTS==2.5.4
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.1.0
HyperPyYAML==1.2.3
idna==3.11
isort==7.0.0
jaconv==0.4.1
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==0.10.0
joblib==1.5.2
kaldiio==2.18.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
modelscope==1.15.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6
numba==0.62.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnxruntime==1.23.2
openai==2.9.0
oss2==2.19.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.1
pooch==1.8.2
postgrest==2.27.2
propcache==0.4.1
protobuf==6.33.2
psycopg2-binary==2.9.11
pyarrow==22.0.0
pyarrow-hotfix==0.7
PyAudio==0.2.14
pycodestyle==2.14.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
pydub==0.25.1
pyflakes==3.4.0
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pylint==4.0.4
pynndescent==0.5.13
pyparsing==3.3.2
pyperclip==1.11.0
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.20
pytorch-wpe==0.0.1
pytz==2025.2
PyYAML==6.0.3
realtime==2.27.2
regex==2025.11.3
requests==2.31.0
rich==14.3.1
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
speechbrain==1.0.3
starlette==0.27.0
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
supafunc==0.10.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.1
torch-complex==0.4.4
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
triton==3.5.1
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
webrtcvad==2.0.10
websocket-client==1.9.0
websockets==13.1
wsproto==1.3.2
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
